// src/containers/shooting_days/enums.ts
export enum ShootingDaysActionTypes {
	SET_SHOOTING_DAY_PREVIEW = 'SET_SHOOTING_DAY_PREVIEW',
	SET_SHOOTING_DAYS = 'SET_SHOOTING_DAYS',
	ADD_NEW_SHOOTING_DAY = 'ADD_NEW_SHOOTING_DAY',
	DELETE_SHOOTING_DAY = 'DELETE_SHOOTING_DAY',
	SET_SHOOTING_DAY_PARAMETER = 'SET_SHOOTING_DAY_PARAMETER',
	DELETE_SHOOTING_DAY_PARAMETER = 'DELETE_SHOOTING_DAY_PARAMETER',
	SET_SHOOTING_DAY_PARAMETER_VALUE = 'SET_SHOOTING_DAY_PARAMETER_VALUE',
	SET_SHOOTING_DAY_SCENE_PARAMETER = 'SET_SHOOTING_DAY_SCENE_PARAMETER',
	SET_POST_SHOOTING_DAY_VALUE = 'SET_POST_SHOOTING_DAY_VALUE',
	SET_ACTIVE_DAY = 'SET_ACTIVE_DAY',
	UPDATE_SCENE_TASKS = 'UPDATE_SCENE_TASKS',
	UPDATE_SCENE_TASKS_INFO = 'UPDATE_SCENE_TASKS_INFO',
	REMOVE_SCENE_TASK = 'REMOVE_SCENE_TASK',
	UPDATE_SCENE_OTHER_TASKS = 'UPDATE_SCENE_OTHER_TASKS',
	ADD_SCENE_OTHER_TASKS_CATEGORY = 'ADD_SCENE_OTHER_TASKS_CATEGORY',
	ADD_SCENE_OTHER_TASK = 'ADD_SCENE_OTHER_TASK',
	REMOVE_SCENE_OTHER_TASK = 'REMOVE_SCENE_OTHER_TASK',
	REMOVE_SCENE_OTHER_TASKS_CATEGORY = 'REMOVE_SCENE_OTHER_TASKS_CATEGORY',
	ADD_SCENE_CHARACTER = 'ADD_SHOOTING_DAY_CHARACTER',
	REMOVE_SCENE_CHARACTER = 'REMOVE_SCENE_CHARACTER',
	REMOVE_SHOOTING_DAY_CHARACTER = 'REMOVE_SHOOTING_DAY_CHARACTER',
	DELETE_SHOOTING_DAY_SCENE = 'DELETE_SHOOTING_DAY_SCENE',
	SET_SHOOTING_DAY_SUPPLIRES = 'SET_SHOOTING_DAY_SUPPLIRES',
	DELETE_SHOOTING_DAY_SUPPLIRES = 'DELETE_SHOOTING_DAY_SUPPLIRES',
	DELETE_SHOOTING_DAY_EMPLOYEES = 'DELETE_SHOOTING_DAY_EMPLOYEES',
	SET_SHOOTING_DAY_EMPLOYEES = 'SET_SHOOTING_DAY_EMPLOYEES',
	SET_SHOOTING_DAY_CHARACTERS = 'SET_SHOOTING_DAY_CHARACTERS',
	UPDATE_SHOOTING_DAY_ACTORS = 'UPDATE_SHOOTING_DAY_ACTORS',
	UPDATE_SHOOTING_DAY_ACTOR_PARAMS = 'UPDATE_SHOOTING_DAY_ACTOR_PARAMS',
	UPDATE_SHOOTING_DAY_EMPLOYEE_PARAMS = 'UPDATE_SHOOTING_DAY_EMPLOYEE_PARAMS',
	ADD_SCENE_TASKS_INFO_ROW = 'ADD_SCENE_TASKS_INFO_ROW',
	DELETE_SCENE_TASKS_INFO_ROW = 'DELETE_SCENE_TASKS_INFO_ROW',
	UPDATE_SHOOTING_DAY_TASK_COST = 'UPDATE_SHOOTING_DAY_TASK_COST',

	UPDATE_SCENE_TAKE = 'UPDATE_SCENE_INFO',
	ADD_SCENE_TAKE = 'ADD_SCENE_INFO_ROW',
	DELETE_SCENE_TAKE = 'DELETE_SCENE_INFO_ROW'
}